Split sites/index into list.ts + lifecycle.ts (one op per file)
index.ts doubled as a topic file holding listSites/startSite/stopSite rather than a package root, so importing @studio/common/sites gave only those three ops. Now each site op lives in its own concern file: listSites -> list.ts (used by the desktop site-server), startSite/stopSite -> lifecycle.ts (CLI one-shot start/stop, for the studio ui server — the desktop runs sites via its own long-lived SiteServer). No re-export barrel, matching the repo's deep-import convention and avoiding eager-loading heavy deps.
