+++ tools/common/sites/list.ts
@@ -2,11 +2,6 @@ import { z } from 'zod';
 import { siteListItemSchema, type SiteListItem } from '@studio/common/lib/cli-events';
 import type { ExecuteCliCommand } from '@studio/common/lib/cli-process';
 
-/**
- * Site operations, delegated to the Studio CLI. Each caller passes its
- * {@link ExecuteCliCommand}, which knows the CLI binary to fork.
- */
-
 // The CLI's `site list --format json` reports the array over its IPC channel as
 // a `keyValuePair` ("sites" → JSON string), the same envelope the desktop reads.
 const siteListKeyValueSchema = z.object( {
@@ -18,6 +13,7 @@ const siteListKeyValueSchema = z.object( {
 		.pipe( z.array( siteListItemSchema ) ),
 } );
 
+/** List the user's local sites via the Studio CLI. */
 export function listSites( execute: ExecuteCliCommand ): Promise< SiteListItem[] > {
 	return new Promise( ( resolve, reject ) => {
 		const [ emitter ] = execute( [ 'site', 'list', '--format', 'json' ], {
@@ -34,35 +30,6 @@ export function listSites( execute: ExecuteCliCommand ): Promise< SiteListItem[]
 		// No `keyValuePair` arrived before exit — treat as an empty list.
 		emitter.on( 'success', () => resolve( [] ) );
 		emitter.on( 'failure', ( { error } ) => reject( error ) );
-		emitter.on( 'error', ( { error } ) => reject( error ) );
-	} );
-}
-
-export function startSite( execute: ExecuteCliCommand, sitePath: string ): Promise< void > {
-	return new Promise( ( resolve, reject ) => {
-		const [ emitter ] = execute(
-			[ 'site', 'start', '--path', sitePath, '--skip-browser', '--skip-log-details' ],
-			{ output: 'capture' }
-		);
-		emitter.on( 'success', () => resolve() );
-		emitter.on( 'failure', ( { error } ) => {
-			error.baseMessage = 'Failed to start site';
-			reject( error );
-		} );
-		emitter.on( 'error', ( { error } ) => reject( error ) );
-	} );
-}
-
-export function stopSite( execute: ExecuteCliCommand, sitePath: string ): Promise< void > {
-	return new Promise( ( resolve, reject ) => {
-		const [ emitter ] = execute( [ 'site', 'stop', '--path', sitePath ], {
-			output: 'capture',
-		} );
-		emitter.on( 'success', () => resolve() );
-		emitter.on( 'failure', ( { error } ) => {
-			error.baseMessage = 'Failed to stop site';
-			reject( error );
-		} );
 		emitter.on( 'error', ( { error } ) => reject( error ) );
 	} );
 }
